Refine scaffold implementation details

+++ backend/app/models/sale.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from datetime import datetime
 
 from sqlalchemy import DateTime, ForeignKey, Numeric, String
@@ -12,7 +13,7 @@ class SaleRecord(Base):
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     card_id: Mapped[int] = mapped_column(ForeignKey("cards.id", ondelete="CASCADE"), index=True)
     sale_date: Mapped[datetime] = mapped_column(DateTime(timezone=True), index=True)
-    sale_price: Mapped[float] = mapped_column(Numeric(12, 2))
+    sale_price: Mapped[Decimal] = mapped_column(Numeric(12, 2))
     currency: Mapped[str] = mapped_column(String(8), default="USD")
     source_name: Mapped[str] = mapped_column(String(120), default="synthetic")
